requirements.txt
setuptools>=39.2.0
wheel
twine
requests
nose
privex-helpers[setuppy]>=2.18.0

